Hopefully fixed #3 and fixed #4.
The fix for #3 is probably not the best method. It uses a cleanup
routine that is called every 10000 packets (or set with -r option).

The fix for #4 is a simple check on the validity of the flow. I also had
to fix one of the items on the TODO list: Check for UDP packets in each
direction.