Major reconstruction of feature storage in flow
All features are now stored in a dictionary, and the feature names can
be accessed through the /features/ list. This tidy's up the code
significantly.

The same should be applied to count variables (those that
are prefixed with c_).

The export (more directly, the str) function can still use some
cleaning up due to this change. Flowtbag needs to be modified to deal
with timestamps as integers, and then the export function can just run
through the list of features and dump them out.